Update to how tags are named for Circonus tagging.
Made a couple of changes to how tags are created - use better names for
the tags and prepend the tags with "aws" to make it clear that they are
AWS tags. Update the test to find user-created vs cmi-tool created tags
to look for tags starting with either "ec2-" or "aws-".